Struct nsstring::nsAString

#[repr(C)]

pub struct nsAString { /* fields omitted */ }

This type is the abstract type which is used for interacting with strings in rust. Each string type can derefence to an instance of this type, which provides the useful operations on strings.

NOTE: Rust thinks this type has a size of 0, because the data associated with it is not necessarially safe to move. It is not safe to construct a nsAString yourself, unless it is received by dereferencing one of these types.

NOTE: The [u8; 0] member is zero sized, and only exists to prevent the construction by code outside of this module. It is used instead of a private () member because the improper_ctypes lint complains about some ZST members in extern "C" function declarations.

Methods

impl nsAString

pub fn assign<T: nsStringLike + ?Sized>(&mut self, other: &T)

Assign the value of other into self, overwriting any value currently stored. Performs an optimized assignment when possible if other is a nsA[C]String.

pub fn fallible_assign<T: nsStringLike + ?Sized>(
&mut self,
other: &T
) -> Result<(), ()>

Assign the value of other into self, overwriting any value currently stored. Performs an optimized assignment when possible if other is a nsA[C]String.

Returns Ok(()) on success, and Err(()) if the allocation failed.

pub fn take_from(&mut self, other: &mut nsAString)

Take the value of other and set self, overwriting any value currently stored. The passed-in string will be truncated.

pub fn fallible_take_from(&mut self, other: &mut nsAString) -> Result<(), ()>

Take the value of other and set self, overwriting any value currently stored. If this function fails, the source string will be left untouched, otherwise it will be truncated.

Returns Ok(()) on success, and Err(()) if the allocation failed.

pub fn append<T: nsStringLike + ?Sized>(&mut self, other: &T)

Append the value of other into self.

pub fn fallible_append<T: nsStringLike + ?Sized>(
&mut self,
other: &T
) -> Result<(), ()>

Append the value of other into self.

Returns Ok(()) on success, and Err(()) if the allocation failed.

pub unsafe fn set_length(&mut self, len: u32)

Set the length of the string to the passed-in length, and expand the backing capacity to match. This method is unsafe as it can expose uninitialized memory when len is greater than the current length of the string.

pub unsafe fn fallible_set_length(&mut self, len: u32) -> Result<(), ()>

Set the length of the string to the passed-in length, and expand the backing capacity to match. This method is unsafe as it can expose uninitialized memory when len is greater than the current length of the string.

Returns Ok(()) on success, and Err(()) if the allocation failed.

pub fn truncate(&mut self)

pub fn to_mut(&mut self) -> &mut [u16]

Get a &mut reference to the backing data for this string. This method will allocate and copy if the current backing buffer is immutable or shared.

pub fn fallible_to_mut(&mut self) -> Result<&mut [u16], ()>

Get a &mut reference to the backing data for this string. This method will allocate and copy if the current backing buffer is immutable or shared.

Returns Ok(&mut [T]) on success, and Err(()) if the allocation failed.

impl nsAString

pub fn assign_utf8<T: nsCStringLike + ?Sized>(&mut self, other: &T)

pub fn fallible_assign_utf8<T: nsCStringLike + ?Sized>(
&mut self,
other: &T
) -> Result<(), ()>

pub fn append_utf8<T: nsCStringLike + ?Sized>(&mut self, other: &T)

pub fn fallible_append_utf8<T: nsCStringLike + ?Sized>(
&mut self,
other: &T
) -> Result<(), ()>