Module nsstring::test_helpers

This module only exists to help with ensuring that the layout of the structs inside of rust and C++ are identical.

It is public to ensure that these testing functions are avaliable to gtest code.

Functions

Rust_Test_Member_nsCString_mData

Rust_Test_Member_nsCString_mFlags

Rust_Test_Member_nsCString_mLength

Rust_Test_Member_nsFixedCString_mFixedBuf

Rust_Test_Member_nsFixedCString_mFixedCapacity

Rust_Test_Member_nsFixedString_mFixedBuf

Rust_Test_Member_nsFixedString_mFixedCapacity

Rust_Test_Member_nsString_mData

Rust_Test_Member_nsString_mFlags

Rust_Test_Member_nsString_mLength

Rust_Test_NsStringFlags

Rust_Test_ReprSizeAlign_nsCString

Rust_Test_ReprSizeAlign_nsFixedCString

Rust_Test_ReprSizeAlign_nsFixedString

Rust_Test_ReprSizeAlign_nsString