Crate nsstring

This module provides rust bindings for the XPCOM string types.

TL;DR (what types should I use)

Use &{mut,} nsA[C]String for functions in rust which wish to take or mutate XPCOM strings. The other string types Deref to this type.

Use ns[C]String<'a> (ns[C]String in C++) for string struct members, and as an intermediate between rust string data structures (such as String, Vec<u16>, &str, and &[u16]) and &{mut,} nsA[C]String (using ns[C]String::from(value)). These conversions, when possible, will not perform any allocations. When using this type in structs shared with C++, the correct lifetime argument is usually 'static.

Use nsFixed[C]String or ns_auto_[c]string! for dynamic stack allocated strings which are expected to hold short string values.

Use *{const,mut} nsA[C]String ({const,} nsA[C]String* in C++) for function arguments passed across the rust/C++ language boundary.

String Types

nsA[C]String

The core types in this module are nsAString and nsACString. These types are zero-sized as far as rust is concerned, and are safe to pass around behind both references (in rust code), and pointers (in C++ code). They represent a handle to a XPCOM string which holds either u16 or u8 characters respectively. The backing character buffer is guaranteed to live as long as the reference to the nsAString or nsACString.

These types in rust are simply used as dummy types. References to them represent a pointer to the beginning of a variable-sized #[repr(C)] struct which is common between both C++ and Rust implementations. In C++, their corresponding types are also named nsAString or nsACString, and they are defined within the nsTSubstring.{cpp,h} file.

Valid Operations

An &nsA[C]String acts like rust's &str, in that it is a borrowed reference to the backing data. When used as an argument to other functions on &mut nsA[C]String, optimizations can be performed to avoid copying buffers, as information about the backing storage is preserved.

An &mut nsA[C]String acts like rust's &mut Cow<str>, in that it is a mutable reference to a potentially borrowed string, which when modified will ensure that it owns its own backing storage. This type can be appended to with the methods .append, .append_utf{8,16}, and with the write! macro, and can be assigned to with .assign.

ns[C]String<'a>

This type is an maybe-owned string type. It acts similarially to a Cow<[{u8,u16}]>. This type provides Deref and DerefMut implementations to nsA[C]String, which provides the methods for manipulating this type. This type's lifetime parameter, 'a, represents the lifetime of the backing storage. When modified this type may re-allocate in order to ensure that it does not mutate its backing storage.

ns[C]Strings can be constructed either with ns[C]String::new(), which creates an empty ns[C]String<'static>, or through one of the provided From implementations. Both string types may be constructed From<&'a str>, with nsCString having a 'a lifetime, as the storage is shared with the str, while nsString has a 'static lifetime, as its storage has to be transcoded.

When passing this type by reference, prefer passing a &nsA[C]String or &mut nsA[C]String. to passing this type.

When passing this type across the language boundary, pass it as *const nsA[C]String for an immutable reference, or *mut nsA[C]String for a mutable reference. This struct may also be included in #[repr(C)] structs shared with C++.

nsFixed[C]String<'a>

This type is a string type with fixed backing storage. It is created with nsFixed[C]String::new(buffer), passing a mutable reference to a buffer as the argument. This buffer will be used as backing storage whenever the resulting string will fit within it, falling back to heap allocations only when the string size exceeds that of the backing buffer.

Like ns[C]String, this type dereferences to nsA[C]String which provides the methods for manipulating the type, and is not #[repr(C)].

When passing this type by reference, prefer passing a &nsA[C]String or &mut nsA[C]String. to passing this type.

When passing this type across the language boundary, pass it as *const nsA[C]String for an immutable reference, or *mut nsA[C]String for a mutable reference. This struct may also be included in #[repr(C)] structs shared with C++, although nsFixed[C]String objects are uncommon as struct members.

ns_auto_[c]string!($name)

This is a helper macro which defines a fixed size, (currently 64 character), backing array on the stack, and defines a local variable with name $name which is a nsFixed[C]String using this buffer as its backing storage.

Usage of this macro is similar to the C++ type nsAuto[C]String, but could not be implemented as a basic type due to the differences between rust and C++'s move semantics.

ns[C]StringRepr

This crate also provides the type ns[C]StringRepr which acts conceptually similar to an ns[C]String<'static>, however, it does not have a Drop implementation.

If this type is dropped in rust, it will not free its backing storage. This can be useful when implementing FFI types which contain ns[C]String members which invoke their member's destructors through C++ code.

Modules

test_helpers
This module only exists to help with ensuring that the layout of the structs inside of rust and C++ are identical.

Macros

ns_auto_cstring

ns_auto_string

Structs

nsACString
This type is the abstract type which is used for interacting with strings in rust. Each string type can derefence to an instance of this type, which provides the useful operations on strings.

nsAString
This type is the abstract type which is used for interacting with strings in rust. Each string type can derefence to an instance of this type, which provides the useful operations on strings.

nsCString

nsCStringRepr
The representation of a ns[C]String type in C++. This type is used internally by our definition of ns[C]String to ensure layout compatibility with the C++ ns[C]String type.

nsFixedCString
A nsFixed[C]String is a string which uses a fixed size mutable backing buffer for storing strings which will fit within that buffer, rather than using heap allocations.

nsFixedString
A nsFixed[C]String is a string which uses a fixed size mutable backing buffer for storing strings which will fit within that buffer, rather than using heap allocations.

nsString

nsStringRepr
The representation of a ns[C]String type in C++. This type is used internally by our definition of ns[C]String to ensure layout compatibility with the C++ ns[C]String type.